Extract added-field body-shape scans from TransformWorkerProgram
Move the listed members out of TransformWorkerProgram without changing their bodies.
